ast_pool: Fix `has_flag` to treat `AstFlag` as `u16` instead of `u8`
This issue previously led to all with a values exceeding the size of a `u8`
being ignored.

+++ core/ast_pool.cpp
@@ -266,7 +266,7 @@ bool has_next_sibling(const AstNode* node) noexcept
 
 bool has_flag(const AstNode* node, AstFlag flag) noexcept
 {
-	return (static_cast<u8>(node->flags) & static_cast<u8>(flag)) != 0;
+	return (static_cast<u16>(node->flags) & static_cast<u16>(flag)) != 0;
 }
 
 AstNode* next_sibling_of(AstNode* node) noexcept
